pyproject.toml
[project]
name = "eden"
version = "0.1.0"
description = "Estimation of Distribution using ENergy-based models"
authors = [
    { name = "Francesco Gentile", email = "francesco.gentile@studenti.unitn.it" },
    { name = "Manuela Corte Pause", email = "manuela.cortepause@studenti.unit.it" },
]
dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
    "entmax>=1.3",
    "wandb",
    "tqdm",
    "coolname",
    "pyyaml",
    "pyserde>=0.17"
]
readme = "README.md"
license.file = "LICENSE"
requires-python = ">= 3.10"

[project.scripts]
eden = "eden:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
